+++ llpc/context/llpcCompiler.cpp
@@ -2169,4 +2169,28 @@ void Compiler::LinkRelocatableShaderElf(
     writer.WriteToBuffer(pPipelineElf);
 }
 
+// =====================================================================================================================
+// Convert front-end LLPC shader stage to middle-end LGC shader type
+lgc::ShaderStage GetLgcShaderStage(Llpc::ShaderStage stage)
+{
+    switch (stage)
+    {
+    case ShaderStageCompute:
+        return lgc::ShaderStageCompute;
+    case ShaderStageVertex:
+        return lgc::ShaderStageVertex;
+    case ShaderStageTessControl:
+        return lgc::ShaderStageTessControl;
+    case ShaderStageTessEval:
+        return lgc::ShaderStageTessEval;
+    case ShaderStageGeometry:
+        return lgc::ShaderStageGeometry;
+    case ShaderStageFragment:
+        return lgc::ShaderStageFragment;
+    default:
+        llvm_unreachable("");
+        return lgc::ShaderStageInvalid;
+    }
+}
+
 } // Llpc
